+++ sdk/identity/azure-identity/azure/identity/_constants.py
@@ -7,6 +7,13 @@
 AZURE_CLI_CLIENT_ID = "04b07795-8ddb-461a-bbee-02f9e1bf7b46"
 
 
+class KnownAuthorities:
+    AZURE_CHINA = "login.chinacloudapi.cn"
+    AZURE_GERMANY = "login.microsoftonline.de"
+    AZURE_GOVERNMENT = "login.microsoftonline.us"
+    AZURE_PUBLIC_CLOUD = "login.microsoftonline.com"
+
+
 class EnvironmentVariables:
     AZURE_CLIENT_ID = "AZURE_CLIENT_ID"
     AZURE_CLIENT_SECRET = "AZURE_CLIENT_SECRET"
@@ -28,5 +35,4 @@ class Endpoints:
     # https://docs.microsoft.com/en-us/azure/active-directory/managed-identities-azure-resources/how-to-use-vm-token#get-a-token-using-http
     IMDS = "http://169.254.169.254/metadata/identity/oauth2/token"
 
-    # TODO: other clouds have other endpoints
-    AAD_OAUTH2_V2_FORMAT = "https://login.microsoftonline.com/{}/oauth2/v2.0/token"
+    AAD_OAUTH2_V2_FORMAT = "https://" + KnownAuthorities.AZURE_PUBLIC_CLOUD + "/{}/oauth2/v2.0/token"