+++ sdk/keyvault/azure-keyvault-certificates/samples/hello_world_async.py
@@ -62,9 +62,12 @@ async def run_sample():
                                         san_dns_names=['sdk.azure-int.net']
                                         )
         cert_name = "HelloWorldCertificate"
+
+        # create_certificate returns a poller. Awaiting the poller will return the certificate
+        # if creation is successful, and the CertificateOperation if not.
         create_certificate_poller = await client.create_certificate(name=cert_name, policy=cert_policy)
-        await create_certificate_poller
-        print("Certificate with name '{0}' created".format(cert_name))
+        certificate = await create_certificate_poller
+        print("Certificate with name '{0}' created".format(certificate.name))
 
         # Let's get the bank certificate using its name
         print("\n.. Get a Certificate by name")
@@ -74,14 +77,14 @@ async def run_sample():
         # After one year, the bank account is still active, and we have decided to update the tags.
         print("\n.. Update a Certificate by name")
         tags = {"a": "b"}
-        updated_certificate = await client.update_certificate(name=bank_certificate.name, tags=tags)
+        updated_certificate = await client.update_certificate_properties(name=bank_certificate.name, tags=tags)
         print("Certificate with name '{0}' was updated on date '{1}'".format(
             bank_certificate.name,
-            updated_certificate.updated)
+            updated_certificate.properties.updated)
         )
         print("Certificate with name '{0}' was updated with tags '{1}'".format(
             bank_certificate.name,
-            updated_certificate.tags)
+            updated_certificate.properties.tags)
         )
 
         # The bank account was closed, need to delete its credentials from the Key Vault.