+++ sdk/storage/azure-storage-blob/azure/storage/blob/aio/blob_service_client_async.py
@@ -392,6 +392,7 @@ class BlobServiceClient(AsyncStorageAccountHostsMixin, BlobServiceClientBase):
                 :caption: Creating a container in the blob service.
         """
         container = self.get_container_client(name)
+        kwargs.setdefault('merge_span', True)
         await container.create_container(
             metadata=metadata, public_access=public_access, timeout=timeout, **kwargs)
         return container
@@ -451,6 +452,7 @@ class BlobServiceClient(AsyncStorageAccountHostsMixin, BlobServiceClientBase):
                 :caption: Deleting a container in the blob service.
         """
         container = self.get_container_client(container) # type: ignore
+        kwargs.setdefault('merge_span', True)
         await container.delete_container( # type: ignore
             lease=lease,
             timeout=timeout,